Do assets/* routing in router package
