+++ web/files/files_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/cozy/cozy-stack/config"
 	"github.com/cozy/cozy-stack/instance"
 	"github.com/cozy/cozy-stack/vfs"
-	"github.com/cozy/cozy-stack/web"
+	"github.com/cozy/cozy-stack/web/errors"
 	"github.com/labstack/echo"
 	"github.com/spf13/afero"
 	"github.com/stretchr/testify/assert"
@@ -1011,7 +1011,7 @@ func TestMain(m *testing.M) {
 	}
 
 	handler := echo.New()
-	handler.HTTPErrorHandler = web.ErrorHandler
+	handler.HTTPErrorHandler = errors.ErrorHandler
 	handler.Use(injectInstance(testInstance))
 	Routes(handler.Group("/files"))
